requirements.txt
pandas==2.2.2
requests==2.32.3
discord==2.3.2
python-dotenv==1.0.1
tabulate==0.9.0
aiohttp==3.9.5
aiocron==1.8